python-run-dataflow-flex-template/requirements.txt
google-api-python-client==2.3.0
google-auth==1.30.0
